+++ conformance/tests/httproute-timeout-backend-request.yaml
@@ -2,11 +2,16 @@ apiVersion: gateway.networking.k8s.io/v1beta1
 kind: HTTPRoute
 metadata:
   name: backend-request-timeout
+  namespace: gateway-conformance-infra
 spec:
   parentRefs:
   - name: same-namespace
   rules:
-  - backendRefs:
+  - matches:
+    - path:
+        type: PathPrefix
+        value: /backend-timeout
+    backendRefs:
     - name: infra-backend-delay
       port: 8080
     timeouts:
